feat(sui-decorators): make more relisiant to Redis lost connection

+++ packages/sui-decorators/test/server/cacheSpec.js
@@ -140,14 +140,23 @@ describe('Cache in the server', () => {
     const useRedis = process.env.USE_REDIS_IN_SUI_DECORATORS_CACHE
     let redis
     if (useRedis) {
-      redis = require('redis').createClient(6379, 'localhost')
+      try {
+        redis = require('redis').createClient(6379, 'localhost')
+        redis.on('error', () => {
+          redis = null
+        })
+      } catch (err) {}
     } else {
       redis = require('redis-mock').createClient()
     }
 
     beforeEach(done => {
+      if (!redis) {
+        return done()
+      }
       redis.flushdb(() => done())
     })
+
     it('should apply cache for ok simple random number response and not apply cache for inlineError decorated error response', async () => {
       let shouldReturnError = true
